Preserve const quality in plugins.
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@22752 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ plugins/plugin_utils/plugin_files.cpp
@@ -294,22 +294,26 @@ std::string concatPaths ( const char* path1, const char* path2 )
 
 std::string getFileDir(const char* file)
 {
-  std::string f = convertPathToDelims(file);
+  char *f = strdup(convertPathToDelims(file).c_str());
+  if (!f)
+    return std::string();
 
-  char *p = (char*)strrchr(f.c_str(),_DirDelim);
+  char *p = strrchr(f,_DirDelim);
   if (p)
     // it's ok to go one past, cus even if it's the end, that's the
     // NUL char so we can set it to NUL again with out worry
     *(p+1) = 0;
 
-  return std::string(f.c_str());
+  std::string ret = f;
+  free(f);
+  return ret;
 }
 
 std::string getFileExtension(const char* file)
 {
   std::string f = convertPathToDelims(file);
 
-  char *p = (char*)strrchr(f.c_str(),'.');
+  const char *p = strrchr(f.c_str(),'.');
   if (!p)
     // it's ok to go one past, cus even if it's the end, that's the
     // NUL char so we can set it to NUL again with out worry
@@ -320,20 +324,24 @@ std::string getFileExtension(const char* file)
 
 std::string getFileTitle(const char* file)
 {
-  std::string f = convertPathToDelims(file);
-  std::string temp = f;
+  char *f = strdup(convertPathToDelims(file).c_str());
+  if (!f)
+    return std::string();
+  char *temp = f;
 
-  const char *p = strrchr(f.c_str(),_DirDelim);
+  char *p = strrchr(f,_DirDelim);
   if (p)
     // it's ok to go one past, cus even if it's the end, that's the
     // NUL char so we can set it to NUL again with out worry
     temp = p+1;
 
-  char *p2 = (char*)strrchr(temp.c_str(),'.');
+  char *p2 = strrchr(temp,'.');
   if (p2)
     *(p2) = 0; // kill the . too
 
-  return std::string(temp.c_str());
+  std::string ret = temp;
+  free(f);
+  return ret;
 }
 
 std::string getFileText(const char* file)
